+++ javascript/react/src/components/verbs/renderers/Derive/Derive.base.tsx
@@ -13,7 +13,7 @@ import {
 	enumDropdown,
 	FormInputType,
 	VerbForm,
-} from '../../verbForm/index.js'
+} from '../forms/index.js'
 
 /**
  * Provides inputs for a Binarize step.
