+++ javascript/react/src/components/verbs/renderers/Fill/Fill.base.tsx
@@ -6,11 +6,7 @@ import type { FillArgs } from '@datashaper/schema'
 import { memo, useMemo } from 'react'
 
 import type { StepComponentProps } from '../../../../types.js'
-import {
-	type FormInput,
-	FormInputType,
-	VerbForm,
-} from '../../verbForm/index.js'
+import { type FormInput, FormInputType, VerbForm } from '../forms/index.js'
 
 /**
  * Provides inputs for a Fill step.
