+++ javascript/react/src/components/verbs/renderers/Impute/Impute.base.tsx
@@ -6,11 +6,7 @@ import type { ImputeArgs } from '@datashaper/schema'
 import { memo, useMemo } from 'react'
 
 import type { StepComponentBaseProps } from '../../../../types.js'
-import {
-	type FormInput,
-	FormInputType,
-	VerbForm,
-} from '../../verbForm/index.js'
+import { type FormInput, FormInputType, VerbForm } from '../forms/index.js'
 
 /**
  * Just the to/value inputs for an impute.
