+++ javascript/react/src/components/verbs/renderers/Join/Join.base.tsx
@@ -15,7 +15,7 @@ import {
 	joinInputs,
 	tableDropdown,
 	VerbForm,
-} from '../../verbForm/index.js'
+} from '../forms/index.js'
 
 /**
  * Provides inputs for a Join step.
