+++ javascript/react/src/components/verbs/renderers/Spread/Spread.base.tsx
@@ -6,11 +6,7 @@ import type { SpreadArgs } from '@datashaper/schema'
 import { memo, useMemo } from 'react'
 
 import type { StepComponentBaseProps } from '../../../../types.js'
-import {
-	type FormInput,
-	FormInputType,
-	VerbForm,
-} from '../../verbForm/index.js'
+import { type FormInput, FormInputType, VerbForm } from '../forms/index.js'
 
 /**
  * Provides inputs for a Spread step.
