+++ javascript/react/src/components/verbs/renderers/Window/Window.base.tsx
@@ -7,7 +7,7 @@ import { WindowFunction } from '@datashaper/schema'
 import { memo, useMemo } from 'react'
 
 import type { StepComponentProps } from '../../../../types.js'
-import { type FormInput, enumDropdown, VerbForm } from '../../verbForm/index.js'
+import { type FormInput, enumDropdown, VerbForm } from '../forms/index.js'
 
 /**
  * Just the column/op inputs for an rollup.
