+++ javascript/react/src/components/verbs/renderers/forms/MultiChoiceInput.tsx
@@ -9,9 +9,9 @@ import { Fragment, memo } from 'react'
 import {
 	useDropdownChangeAllHandler,
 	useDropdownChangeHandler,
-} from '../../../hooks/index.js'
-import type { StepChangeFunction } from '../../../types.js'
-import { dropdownStyles } from '../../styles.js'
+} from '../../../../hooks/index.js'
+import type { StepChangeFunction } from '../../../../types.js'
+import { dropdownStyles } from '../../../styles.js'
 import type { MultiChoiceFormInput } from './MultiChoiceInput.types.js'
 
 export const MultiChoiceInput: React.FC<{
