+++ javascript/react/src/components/verbs/renderers/forms/MultiChoiceInput.types.ts
