+++ javascript/react/src/components/verbs/renderers/shared/DataTypeField.tsx
@@ -8,8 +8,8 @@ import type { IDropdownOption } from '@fluentui/react'
 import { Dropdown, SpinButton } from '@fluentui/react'
 import { memo, useCallback } from 'react'
 
-import { CalendarPicker } from '../../controls/index.js'
-import { narrowerDropdownStyles } from '../../styles.js'
+import { CalendarPicker } from '../../../controls/index.js'
+import { narrowerDropdownStyles } from '../../../styles.js'
 import { Container, TextValue } from './DataTypeField.styles.js'
 import type { DataTypeFieldProps } from './DataTypeField.types.js'
