+++ javascript/react/src/components/verbs/renderers/shared/FilterFunction.types.ts
